feat: persist deps-stats package across navigation
